+++ src/components/Label/Label.tsx
@@ -14,6 +14,9 @@ interface LabelProps {
   /** Controls whether the "(Required)" text should be displayed alongside the
    * label's text. False by default. */
   isRequired?: boolean;
+  /** Allows the '(Required)' text to be changed for language purposes
+   * Note: Parenthesis will be added automatically by the component */
+  requiredLabelText?: string;
 }
 
 /**
@@ -31,6 +34,7 @@ export const Label = chakra(
       id,
       isInlined = false,
       isRequired = false,
+      requiredLabelText,
       ...rest
     } = props;
     const styles = useStyleConfig("Label", { isInlined });
@@ -52,7 +56,11 @@ export const Label = chakra(
         {...rest}
       >
         {children}
-        {isRequired && <span> (Required)</span>}
+        {isRequired && (
+          <span>
+            {requiredLabelText ? ` (${requiredLabelText})` : " (Required)"}
+          </span>
+        )}
       </Box>
     );
   })